# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop-api Documentation",
        "description": "petshop api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/colaboradores": {
            "get": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Listar todos os Colaboradores",
                "description": "Retorna uma lista de todos os Colaboradores",
                "operationId": "21e559a3a9d3a89d04a71175932a691d",
                "responses": {
                    "200": {
                        "description": "Lista de Colaboradores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Colaborador"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Criar um novo Colaborador",
                "description": "Cria um novo Colaborador com base nos dados fornecidos",
                "operationId": "3f9de1a07b7d9be7877217e7d3b95596",
                "requestBody": {
                    "description": "Dados do novo Colaborador",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Colaborador"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Colaborador criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Colaborador"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/colaboradores/{id}": {
            "get": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Obter detalhes de um Colaborador",
                "description": "Retorna os detalhes de um Colaborador com base no ID",
                "operationId": "83915c2869e235bab54ccd5d1c8a4501",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do Colaborador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Colaborador"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Colaborador não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Colaborador não encontrado"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Atualizar um Colaborador",
                "description": "Atualiza um Colaborador com base no ID e nos dados fornecidos",
                "operationId": "cbc0ce2f4cf51430c9e01ed1c8b962d6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados atualizados do Colaborador",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Colaborador"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Colaborador atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Colaborador"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Colaborador não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Colaborador não encontrado"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Colaboradores"
                ],
                "summary": "Excluir um Colaborador",
                "description": "Exclui um Colaborador com base no ID",
                "operationId": "ead759d963368ffbbc6a4401d4f6da0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Colaborador",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Colaborador excluído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Colaborador excluído com sucesso"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Colaborador não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "Colaborador não encontrado"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Colaborador": {
                "title": "Colaborador",
                "description": "Modelo de dados para um Colaborador",
                "properties": {
                    "id": {
                        "description": "ID do Colaborador",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nome": {
                        "description": "Nome do Colaborador",
                        "type": "string"
                    },
                    "cpf": {
                        "description": "CPF do Colaborador",
                        "type": "string"
                    },
                    "email": {
                        "description": "E-mail do Colaborador",
                        "type": "string"
                    },
                    "unidade_id": {
                        "description": "ID da Unidade do Colaborador",
                        "type": "integer",
                        "format": "int64"
                    },
                    "cargo_id": {
                        "description": "ID do Cargo do Colaborador",
                        "type": "integer",
                        "format": "int64"
                    },
                    "nota_desempenho": {
                        "description": "Nota de desempenho do Colaborador",
                        "type": "number",
                        "format": "float"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}